Fixed bug when assignments YAML file is missing.
As part of this, we've extracted a class to be responsible for holding current assignments, which can also deal with persisting them.

We've also extracted the fairness tests of the `Administrator` class from those that demonstrate it's API (which now really just call the corresponding methods on the `Assignments` store), and dramatically simplified the method which chooses an assignee fairly.