An attempt at fixing #10.
- The very reasonable supposition is that since each Harmonia assign,
unassign & remind in the schedule is converted into a line in the
crontab, multiple processes will be trying to write to the
assignments.yml file at the same time. This is particularly likely at
midday on Monday when all tasks are being unassigned at the same time.
- I haven't attempted to reproduce the problem and I haven't attempted
to fix the underlying problem i.e. the non-threadsafe writing to
assignments.yml. The simplest way to do that might be to start using a
database, but that seems overkill at the moment.
- What I've attempted to do is to modify Harmonia so it's possible to
supply a list of tasks to assign, unassign or remind, and then adapted
the schedule to make use of this new functionality such that each
operation is done in one line of the crontab and we should not end up
with multiple processes trying to write to the assignments.yml file at
the same time.
- The syntax allowed by whenever is quite restricted and particularly
awkward for us, because it uses the Ruby symbol syntax for interpolation
of arguments. I don't think the change I've made to the schedule makes
it much less readable, but it's not great.
- The tests I've added for my changes to the Harmonia class aren't
great, but hopefully they're better than nothing.