Attempt to fix #4.
- I think I've tracked this down to the pound sign in the erb template.
It seems like an acceptable fix for now to use GPB instead of a pound
sign.
- Since it's not something coming back from FreeAgent, I've removed the
UTF-8 conversions that look like they've been added with the suspicion
that the nasty character was coming back from FreeAgent.
- I've re-enabled the FreeAgent invoices query on the basis that I think
it will now work, but I plan to test it again once I've deployed.