Fix test errors when parsing an empty YAML file.
The errors started occurring on Ruby 1.9.3-p194 which is what Travis CI
is running. On Travis CI, the error was Psych::SyntaxError "control
characters are not allowed at line 1 column 1".

However, when I installed the same version of Ruby locally, I got
errors in the same two tests and raised from the same line of code,
but the error was TypeError: "no implicit conversion from nil to
integer". I suspect the discrepancy is because I'm on OSX and Travis CI
is presumably running some flavour of Linux and the YAML library depends
on libyaml which may be slightly different on the two platforms.

Given that we're not actually running this latest version of Ruby in
production, the simplest solution for now is to rescue the exception and
default to an empty Hash for the assignments. Hopefully this will work
on Travis CI too.