Unassign all tasks on Monday morning
Tom has been assigned the drinks task since the 1st of August and James
M has been assigned the gardener task since (according to the emails)
the 19th of July.  The effect of this is that James A and I have been
sharing the week notes and invoices since the beginning of August.