Move the PAYE instructions into their own task
Having spoken to HMRC today I've confirmed that our quarterly payments
are due on the 5th of April, July, October and January.  The earliest we
can pay is the 6th and the payment deadline is the 19th of the same
month.  Given this timing it no longer makes sense to combine the PAYE
payments with the wages task that is triggered on the 25th of each
month.