Upgrade recap to 1.0.5
This fixes some of the problems that still existed in the version of
recap we were tied to, and changes the deploy path from
/home/harmonia/apps/harmonia to /home/harmonia/app.

I've already deployed this version to our server with the following
commands:

    $ cap bootstrap
    $ cap deploy:setup
    $ cap deploy

I've also copied the assignments.yml file from
/home/harmonia/apps/harmonia/config to /home/harmonia/app/config, tar'd
up the old apps directory and removed the apps directory.

The commands in Harmonia's crontab have been updated correctly so
everything should *just work*.

*NOTE* I've chosen to remove config/deploy.rb in favour of keeping
everything in the Capfile, which is what recap's `setup` task does by
default.